backend/flask-app/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
psycopg2-binary==2.9.3
Werkzeug==2.0.3
python-dotenv~=1.0.1
openpyxl~=3.1.2
pillow~=10.3.0
requests==2.26.0

celery~=5.4.0